Move vehicle and license count into App.jsx
